Fix Y2038 timestamp overflow in sonic-pac authentication manager
- Change sessionTime from uint32 to uint64 in authmgrClientInfo_t
- Change lastAuthTime from uint32 to uint64 in authmgrClientInfo_t
- Change sessionTime from uint32 to uint64 in authmgrPortSessionStats_t
- Change sessionTime from uint32 to uint64 in pac_authenticated_clients_oper_table_t
- Change lastAuthTime from uint32 to uint64 in pac_authenticated_clients_oper_table_t

These changes prevent timestamp overflow that would occur on
January 19, 2038 when 32-bit signed timestamps exceed INT32_MAX.

+++ src/sonic-pac/authmgr/common/auth_mgr_exports.h
@@ -258,7 +258,7 @@ typedef struct authmgrPortSessionStats_s
   uint32        sessionPacketsGbRx;
   uint32        sessionPacketsTx;
   uint32        sessionPacketsGbTx;
-  uint32        sessionTime;
+  uint64        sessionTime;  /* Y2038-safe: changed from uint32 */
    uchar8        userName[AUTHMGR_USER_NAME_LEN];
    char8         sessionId[AUTHMGR_SESSION_ID_LEN];
 } authmgrPortSessionStats_t;
